refactor: navigation モジュールを削除
