+++ core/ui/build.gradle.kts
@@ -55,6 +55,7 @@ dependencies {
   implementation(libs.androidxComposeUiToolingPreview)
   implementation(libs.androidxHiltNavigationCompose)
   implementation(libs.androidxLifecycleRuntimeCompose)
+  implementation(libs.androidxNavigationCompose)
   implementation(libs.coilCompose)
   implementation(libs.serialization)
 }
