+++ feature/my-book-list/src/main/java/app/kaito_dogi/mybrary/feature/mybooklist/MyBookListNavigation.kt
@@ -3,7 +3,7 @@ package app.kaito_dogi.mybrary.feature.mybooklist
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import app.kaito_dogi.mybrary.core.domain.model.MyBook
-import app.kaito_dogi.mybrary.core.navigation.MybraryRoute
+import app.kaito_dogi.mybrary.core.ui.navigation.MybraryRoute
 
 fun NavGraphBuilder.myBookListScreen(
   onAdditionClick: () -> Unit,
