+++ feature/send-otp/src/main/java/app/kaito_dogi/mybrary/feature/sendotp/SendOtpNavigation.kt
@@ -2,7 +2,7 @@ package app.kaito_dogi.mybrary.feature.sendotp
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import app.kaito_dogi.mybrary.core.navigation.MybraryRoute
+import app.kaito_dogi.mybrary.core.ui.navigation.MybraryRoute
 
 fun NavGraphBuilder.sendOtpScreen(
   onSendOtpComplete: (email: String, MybraryRoute.VerifyOtp.Page) -> Unit,
