+++ settings.gradle.kts
@@ -23,7 +23,6 @@ include(":core:data")
 include(":core:database")
 include(":core:design-system")
 include(":core:domain")
-include(":core:navigation")
 include(":core:supabase")
 include(":core:ui")
 include(":feature:my-book-detail")
